Refactors Helm chart build and push workflow
Streamlines the Helm chart build and push workflow in GitHub Actions.

This change separates the workflow into distinct `setup` and `build` jobs to improve clarity and efficiency. The `setup` job determines the chart version and prepares the environment. The `build` job then depends on `setup` for its output and handles building, packaging, and pushing the Helm chart to GHCR.

The workflow now generates Kubernetes manifest for tagged releases, which are uploaded as release artifacts. It also tags the `latest` chart version for main branch builds. These changes enhance the reliability and maintainability of the deployment process, and make it easier for users to install tagged releases.

Signed-off-by: Lu Ma <malu17@huawei.com>